+++ plugins/governance/changelog.md
@@ -1,5 +1,4 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- Feat: Now Bifrost supports provider level fallbacks
-- Chore: Dependency upgrades
+- Chore: using core 1.2.1 and framework 1.1.1